requirements.txt
django==1.4
django-grappelli==2.4.0
django-tastypie==0.9.11
django-tastypie-mongoengine==0.2.1
django-extensions